refactoring generate_mixture_constraint and deleting unused functions
