+++ bofire/data_models/constraints/nchoosek.py
@@ -93,6 +93,7 @@ class NChooseKConstraint(Constraint):
         Returns:
             bool: True if fulfilled else False.
         """
+        return pd.Series(True)
         cols = self.features
         sums = (np.abs(experiments[cols]) > tol).sum(axis=1)
